Remove CopyToClipboard and move its usage to ShareButtons component
