+++ projects/js-packages/publicize-components/src/components/share-buttons/share-buttons.tsx
@@ -1,8 +1,8 @@
 import { SocialServiceIcon, Button } from '@automattic/jetpack-components';
+import { CopyToClipboard } from '@automattic/jetpack-components';
 import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
 import { useCallback } from '@wordpress/element';
 import { availableNetworks } from './available-networks';
-import { CopyToClipboard } from './copy-to-clipboard';
 import styles from './styles.module.scss';
 import { ShareButtonProps } from './types';
 import { useShareButtonText } from './useShareButtonText';
@@ -18,9 +18,18 @@ export type ShareButtonsProps = ShareButtonProps;
  * @returns {React.JSX.Element} - Rendered component
  */
 export function ShareButtons( { buttonStyle = 'icon', buttonVariant }: ShareButtonsProps ) {
-	const prepareUrl = useShareButtonText();
+	const prepareText = useShareButtonText();
 
 	const { recordEvent } = useAnalytics();
+
+	const onCopy = useCallback( () => {
+		recordEvent( 'jetpack_social_share_button_clicked', { network: 'clipboard' } );
+	}, [ recordEvent ] );
+
+	const textToCopy = useCallback(
+		() => prepareText( '{{text}}\n{{url}}', false ),
+		[ prepareText ]
+	);
 
 	const getOnClick = useCallback(
 		function ( url: string, data?: unknown ) {
@@ -42,7 +51,7 @@ export function ShareButtons( { buttonStyle = 'icon', buttonVariant }: ShareButt
 	return (
 		<div className={ styles[ 'share-buttons' ] }>
 			{ availableNetworks.map( ( { label, networkName, url } ) => {
-				const href = prepareUrl( url );
+				const href = prepareText( url );
 
 				const icon =
 					'text' !== buttonStyle ? <SocialServiceIcon serviceName={ networkName } /> : null;
@@ -64,7 +73,13 @@ export function ShareButtons( { buttonStyle = 'icon', buttonVariant }: ShareButt
 					</Button>
 				);
 			} ) }
-			<CopyToClipboard buttonStyle={ buttonStyle } buttonVariant={ buttonVariant } />
+			<CopyToClipboard
+				buttonStyle={ buttonStyle }
+				onCopy={ onCopy }
+				textToCopy={ textToCopy }
+				className={ styles.clipboard }
+				variant={ buttonVariant }
+			/>
 		</div>
 	);
 }
